fix: category get API 함수 역할 분리하도록 수정
